version1
this program creates two CSVs for two countries based on indicator for a specific month.
it then gets the average percent of whatever the indicator is for each country and puts it in a seperate textfile.
its buggy right now, stops after a single month, but future version will create a text file with an average of each month from may to present that can be then used to make a llinegraph or something